+++ docs/pages/reference/alchemy/aa-infra/functions/alchemyEstimateFeesPerGas.mdx
@@ -1,0 +1,47 @@
+---
+# This file is autogenerated
+title: alchemyEstimateFeesPerGas
+description: Overview of the alchemyEstimateFeesPerGas method
+slug: wallets/reference/alchemy/aa-infra/functions/alchemyEstimateFeesPerGas
+---
+
+A custom `estimateFeesPerGas`function for viem bundler clients to use `rundler_maxPriorityFeePerGas` for priority fee estimation.
+
+It fetches:
+
+1. `baseFeePerGas` from the latest block.
+2. `maxPriorityFeePerGas` via `rundler_maxPriorityFeePerGas`.
+
+It then returns `maxFeePerGas = baseFee * 1.5 + priority` (aligns with viem default).
+
+## Import
+
+```ts
+import { alchemyEstimateFeesPerGas } from "@alchemy/aa-infra";
+```
+
+## Usage
+
+```ts
+import { createBundlerClient } from "viem/account-abstraction";
+import { alchemyEstimateFeesPerGas } from "@alchemy/aa-infra";
+
+const bundler = createBundlerClient({
+  transport: http("<rundler-url>"),
+  userOperation: {
+    estimateFeesPerGas: alchemyEstimateFeesPerGas,
+  },
+});
+```
+
+## Parameters
+
+### bundlerClient
+
+`PriorityFeeClient`
+Bundler client with the rundler RPC method.
+
+## Returns
+
+`Promise<{maxFeePerGas: bigint, maxPriorityFeePerGas: bigint}>`
+Estimated fee values.
