+++ docs/pages/reference/alchemy/aa-infra/functions/createAlchemyBundlerClient.mdx
@@ -1,0 +1,57 @@
+---
+# This file is autogenerated
+title: createAlchemyBundlerClient
+description: Overview of the createAlchemyBundlerClient method
+slug: wallets/reference/alchemy/aa-infra/functions/createAlchemyBundlerClient
+---
+
+Create a viem Bundler Client pre-configured for Alchemy Bundler.
+
+Uses rundler_maxPriorityFeePerGas fee estimation via a custom function.
+Extends the RPC schema so the new method is typed.
+TODO(v5): Revisit the `createAlchemyBundlerClient` export - we might rename or move it in the next major.
+
+## Import
+
+```ts
+import { createAlchemyBundlerClient } from "@alchemy/aa-infra";
+```
+
+## Usage
+
+```ts
+import { createAlchemyBundlerClient } from "@alchemy/aa-infra";
+
+const bundler = createAlchemyBundlerClient({
+  account: smartAccount,
+  chain: mainnet,
+  transport: custom(rpcClient),
+});
+```
+
+## Parameters
+
+### params
+
+`Params`
+The parameter configuration object
+
+### params.account
+
+`SmartAccount`
+The smart account instance for the client
+
+### params.chain
+
+`Chain | undefined`
+The chain the bundler operates on
+
+### params.transport
+
+`Transport`
+The viem transport
+
+## Returns
+
+`BundlerClient`
+A Bundler Client ready to use with Alchemy's Rundler
