+++ packages/smart-accounts/src/light-account/accounts/account.test.ts
@@ -16,7 +16,7 @@ import { singleOwnerLightAccountActions } from "../decorators/singleOwner.js";
 import type { LightAccountVersion } from "../types.js";
 import { AccountVersionRegistry } from "../utils.js";
 import { toLightAccount } from "./account.js";
-import { alchemyEstimateFeesPerGas } from "../../alchemyEstimateFeesPerGas.js";
+import { alchemyEstimateFeesPerGas } from "@alchemy/aa-infra";
 
 const versions = Object.keys(
   AccountVersionRegistry.LightAccount,
