+++ packages/smart-accounts/src/types.ts
@@ -1,9 +1,4 @@
-import type {
-  IsUndefined,
-  SignableMessage,
-  TypedDataDefinition,
-  RpcUserOperation,
-} from "viem";
+import type { IsUndefined, SignableMessage, TypedDataDefinition } from "viem";
 import type { SmartAccount } from "viem/account-abstraction";
 
 export type GetAccountParameter<
@@ -23,12 +18,3 @@ export type SignatureRequest =
       type: "eth_signTypedData_v4";
       data: TypedDataDefinition;
     };
-
-// Minimal RPC schema that only declares the Alchemy priority fee method.
-export type AlchemyRpcSchema = [
-  {
-    Method: "rundler_maxPriorityFeePerGas";
-    Parameters: [];
-    ReturnType: RpcUserOperation["maxPriorityFeePerGas"];
-  },
-];
